Add redirect URL to auth config
It's pretty common for a user to try to access a URL when they're
not logged in. I thought it would be nice if Auth0 would keep
track of this so that I could then use it within my app to redirect
the user to their desired URL upon logging in.